add all progress work

+++ app/Http/Controllers/UserController.php
@@ -31,6 +31,9 @@ class UserController extends Controller
         $validated['date_of_birth'] = date('Y-m-d', strtotime($validated['date_of_birth']));
         $user = auth()->user();
 
+        if ($request->hasFile('avatar')) {
+            return $request->file('avatar');
+        }
         if ($request->file('avatar')) {
             if ($user->avatar) {
                 Storage::delete('public/profile-image/' . $user->avatar);
